Move Local functions to electron only component
