+++ newIDE/app/src/ProjectCreation/LocalStarters.js
@@ -16,7 +16,10 @@ import ListIcon from '../UI/ListIcon';
 import { showGameFileCreationError } from './LocalExamples';
 import { type StorageProvider, type FileMetadata } from '../ProjectsStorage';
 import LocalFileStorageProvider from '../ProjectsStorage/LocalFileStorageProvider';
+import { findEmptyPath } from './LocalPathFinder';
 const path = optionalRequire('path');
+const electron = optionalRequire('electron');
+const app = electron ? electron.remote.app : null;
 var fs = optionalRequire('fs-extra');
 const gd = global.gd;
 
@@ -76,6 +79,16 @@ export default class LocalStarters extends Component<Props, State> {
       fileIdentifier: filePath,
     });
     sendNewGameCreated('');
+  }
+
+  componentDidMount() {
+    if ((this.props.outputPath === ''))
+      if (path && app)
+        this.props.onChangeOutputPath(
+          findEmptyPath(
+            path.join(app.getPath('documents'), 'GDevelop projects')
+          )
+        );
   }
 
   render() {
